# Memoize DescriptorParser (#1875)