@olafurpg 2019-10-31
Closing due to inactivity. We can revisit this in the future if performance of the symbol parsers becomes a bottleneck again.